⚡️ perf: improve layout performance and fix language switch (#1719)
* ⚡️ perf: improve perf

* ⚡️ perf: improve perf

* ♻️ refactor: refactor the layout provider

* ✅ test: fix test

* 🎨 chore: improve code
