+++ src/store/global/slices/common/action.test.ts
@@ -7,11 +7,11 @@ import { globalService } from '@/services/global';
 import { userService } from '@/services/user';
 import { useGlobalStore } from '@/store/global';
 import { GlobalServerConfig } from '@/types/settings';
-import { switchLang } from '@/utils/switchLang';
+import { switchLang } from '@/utils/client/switchLang';
 
 vi.mock('zustand/traditional');
 
-vi.mock('@/utils/switchLang', () => ({
+vi.mock('@/utils/client/switchLang', () => ({
   switchLang: vi.fn(),
 }));
